tasklet: Refactor tasklet code
